# storage/openapi.yml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: "http://localhost:8080"
info:
  description: API for an inventory management system
  version: "1.0.0"
  title: Inventory Management System API
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
paths:
  /getExpense:
    get:
      tags:
        - getExpense
      summary: Get the expenses
      operationId: app.getExpense
      description: Getting the expenses
      parameters:
        - name: start_timestamp
          in: query
          description: Limits the number of items on a page
          schema:
            type: string
            format: date-time
            example: "2016-08-29T09:12:33.001Z"
        - name: end_timestamp
          in: query
          description: Limits the number of items on a page
          schema:
            type: string
            format: date-time
            example: "2016-08-29T09:12:33.001Z"
      responses:
        "200":
          description: Successfully returned a list of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/expenseEvent"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /getRevenue:
    get:
      tags:
        - getRevenue
      summary: Get the revenue reported
      operationId: app.getRevenue
      description: Getting the revenue reported
      parameters:
        - name: start_timestamp
          in: query
          description: Limits the number of items on a page
          schema:
            type: string
            format: date-time
            example: "2016-08-29T09:12:33.001Z"
        - name: end_timestamp
          in: query
          description: Limits the number of items on a page
          schema:
            type: string
            format: date-time
            example: "2016-08-29T09:12:33.001Z"
      responses:
        "200":
          description: Successfully returned a list of revenue reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/reportEvent"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

components:
  schemas:
    expenseEvent:
      type: object
      required:
        - order_id
        - item_id
        - item_name
        - quantity
        - price
        - timestamp
      properties:
        order_id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        item_id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        item_name:
          type: string
          example: Gala Apple
        quantity:
          type: integer
        price:
          type: number
        timestamp:
          format: date-time
          example: "2016-08-29T09:12:33.001Z"
        trace_id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
    reportEvent:
      required:
        - submission_id
        - store_id
        - employee_id
        - revenue
        - report_period
        - timestamp
      properties:
        submission_id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        store_id:
          type: string
          example: VAN-001
        employee_id:
          type: string
        revenue:
          type: number
        report_period:
          type: integer
        timestamp:
          format: date-time
          example: "2016-08-29T09:12:33.001Z"
        trace_id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
      type: object